chore: use a basic script instead
