+++ package.json
@@ -25,6 +25,7 @@
     "test:watch": "jest --watch",
     "lint": "eslint \"./src/**\" \"./scripts/**\" \"./tokens/**\" --no-error-on-unmatched-pattern",
     "release": "semantic-release",
+    "release:deprecations": "node ./scripts/manageDeprecations.mjs deprecations.json",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "chromatic": "npx chromatic --project-token $CHROMATIC_PROJECT_TOKEN --exit-zero-on-changes",
@@ -100,7 +101,6 @@
     "sass": "^1.45.0",
     "sass-loader": "^12.1.0",
     "semantic-release": "^19.0.0",
-    "semantic-release-npm-deprecate-old-versions": "^1.3.4",
     "semver": "^7.7.2",
     "storybook": "^8.6.0-alpha.0",
     "style-dictionary": "^3.9.2",
