+++ scripts/manageDeprecations.mjs
@@ -1,0 +1,23 @@
+/**
+ * This files stores and manages deprecated version ranges.
+ * Whenever critical regressions are found, the `DEPRECATIONS`
+ * list below may be updated with the range of affected versions.
+ *
+ * The release workflow will update the NPM registry with any new
+ * deprecations via `semantic-release`.
+ */
+import { readFileSync } from "fs";
+import { resolve } from "path";
+import semver from "semver";
+import checkVersion from "./util/checkVersion.mjs";
+
+const PATH_PKG = resolve(process.cwd(), "./package.json");
+
+const CLI_ARG = process?.argv[2];
+
+/** Current Major version is the only one with "support" */
+let CONFIG = { versions: [] };
+
+if (CLI_ARG) {
+  CONFIG = JSON.parse(readFileSync(CLI_ARG));
+}
